Remove intermediate editor
