+++ frontend/src/metabase/query_builder/components/NativeQueryEditor/CodeMirrorEditor/CodeMirrorEditor.tsx
@@ -11,8 +11,25 @@ import {
 } from "react";
 
 import { isEventOverElement } from "metabase/lib/dom";
+import type NativeQuery from "metabase-lib/v1/queries/NativeQuery";
+import type { CardId } from "metabase-types/api";
 
-import type { EditorProps, EditorRef } from "../Editor";
+import type { SelectionRange } from "../types";
+
+export type EditorProps = {
+  query: NativeQuery;
+  onChange?: (queryText: string) => void;
+  readOnly?: boolean;
+  onCursorMoveOverCardTag?: (id: CardId) => void;
+  onRightClickSelection?: () => void;
+  onSelectionChange?: (range: SelectionRange) => void;
+};
+
+export interface EditorRef {
+  focus: () => void;
+  resize: () => void;
+  getSelectionTarget: () => Element | null;
+}
 
 import S from "./CodeMirrorEditor.module.css";
 import { useExtensions } from "./extensions";
